+++ packages/article/__tests__/android/article.android.test.js
@@ -9,13 +9,13 @@ jest.mock("@times-components/article-flag", () => ({
   UpdatedArticleFlag: "UpdatedArticleFlag",
   SponsoredArticleFlag: "SponsoredArticleFlag"
 }));
+jest.mock("@times-components/article-image", () => "ArticleImage");
 jest.mock("@times-components/article-label", () => "ArticleLabel");
+jest.mock("@times-components/article-topics", () => "ArticleTopics");
 jest.mock("@times-components/video-label", () => "VideoLabel");
 jest.mock("@times-components/brightcove-video", () => "BrightcoveVideo");
-jest.mock("@times-components/article-image", () => "ArticleImage");
 jest.mock("@times-components/pull-quote", () => "PullQuote");
 jest.mock("@times-components/related-articles", () => "RelatedArticles");
-jest.mock("@times-components/topics", () => "Topics");
 jest.mock("@times-components/watermark", () => "Watermark");
 jest.mock("@times-components/tracking", () => {
   const mockTracking = component => component;