+++ packages/provider/src/provider.js
@@ -8,6 +8,9 @@ import AuthorArticlesWithImagesProvider, {
   query as articleListWithImagesQuery
 } from "./author-articles-with-images";
 import ArticleProvider, { query as articleQuery } from "./article";
+import TopicArticlesProvider, {
+  query as topicArticlesQuery
+} from "./topic-articles";
 
 export default connect;
 
@@ -19,5 +22,7 @@ export {
   articleQuery,
   authorProfileQuery,
   articleListNoImagesQuery,
-  articleListWithImagesQuery
+  articleListWithImagesQuery,
+  TopicArticlesProvider,
+  topicArticlesQuery
 };