+++ packages/date-publication/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.15.7"></a>
+## [0.15.7](https://github.com/newsuk/times-components/compare/@times-components/date-publication@0.15.6...@times-components/date-publication@0.15.7) (2018-05-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/date-publication
+
+<a name="0.15.6"></a>
+## [0.15.6](https://github.com/newsuk/times-components/compare/@times-components/date-publication@0.15.5...@times-components/date-publication@0.15.6) (2018-05-04)
+
+
+### Bug Fixes
+
+* dextrose ([#919](https://github.com/newsuk/times-components/issues/919)) ([a8fa1d4](https://github.com/newsuk/times-components/commit/a8fa1d4))
+
+
+
+
 <a name="0.15.5"></a>
 ## [0.15.5](https://github.com/newsuk/times-components/compare/@times-components/date-publication@0.15.4...@times-components/date-publication@0.15.5) (2018-04-30)