+++ packages/image/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.17.7"></a>
+## [1.17.7](https://github.com/newsuk/times-components/compare/@times-components/image@1.17.6...@times-components/image@1.17.7) (2018-05-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/image
+
+<a name="1.17.6"></a>
+## [1.17.6](https://github.com/newsuk/times-components/compare/@times-components/image@1.17.5...@times-components/image@1.17.6) (2018-05-04)
+
+
+### Bug Fixes
+
+* dextrose ([#919](https://github.com/newsuk/times-components/issues/919)) ([a8fa1d4](https://github.com/newsuk/times-components/commit/a8fa1d4))
+
+
+
+
 <a name="1.17.5"></a>
 ## [1.17.5](https://github.com/newsuk/times-components/compare/@times-components/image@1.17.4...@times-components/image@1.17.5) (2018-05-01)