+++ packages/pagination/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.18.13"></a>
+## [0.18.13](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.18.12...@times-components/pagination@0.18.13) (2018-05-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/pagination
+
+<a name="0.18.12"></a>
+## [0.18.12](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.18.11...@times-components/pagination@0.18.12) (2018-05-04)
+
+
+### Bug Fixes
+
+* dextrose ([#919](https://github.com/newsuk/times-components/issues/919)) ([a8fa1d4](https://github.com/newsuk/times-components/commit/a8fa1d4))
+
+
+
+
 <a name="0.18.11"></a>
 ## [0.18.11](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.18.10...@times-components/pagination@0.18.11) (2018-04-30)