+++ packages/provider/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.29.0"></a>
+# [0.29.0](https://github.com/newsuk/times-components/compare/@times-components/provider@0.28.8...@times-components/provider@0.29.0) (2018-05-08)
+
+
+### Features
+
+* Topic provider ([#926](https://github.com/newsuk/times-components/issues/926)) ([d982278](https://github.com/newsuk/times-components/commit/d982278))
+
+
+
+
+<a name="0.28.8"></a>
+## [0.28.8](https://github.com/newsuk/times-components/compare/@times-components/provider@0.28.7...@times-components/provider@0.28.8) (2018-05-04)
+
+
+### Bug Fixes
+
+* dextrose ([#919](https://github.com/newsuk/times-components/issues/919)) ([a8fa1d4](https://github.com/newsuk/times-components/commit/a8fa1d4))
+
+
+
+
 <a name="0.28.7"></a>
 ## [0.28.7](https://github.com/newsuk/times-components/compare/@times-components/provider@0.28.6...@times-components/provider@0.28.7) (2018-05-03)